refactored commonly used functions to utils.py
