+++ eo-python-runtime/tests/example_programs/test_inline_attributes.py
@@ -5,7 +5,6 @@
 # by the EO compiler v.0.1.25.
 
 from eo2py.atoms import *
-from functools import partial
 
 """
 [n] > fibonacci
@@ -76,15 +75,7 @@ class EOapp(ApplicationMixin, Object):
 
 import pytest
 
-
-def fibonacci(n):
-    if n < 3:
-        return 1
-    else:
-        memo = [1] * n
-        for i in range(2, n):
-            memo[i] = memo[i - 1] + memo[i - 2]
-        return memo[n - 1]
+from .utils import fibonacci
 
 
 @pytest.mark.parametrize("n", list(range(1, 10)))
